squashed multi-intro support with introless mode
This is a big squash merge combining work by several people, including:
 - Faruque Sarker - wrote the original multi-introducer patch in 2010
 - KillYourTV - ported multi-intro patch to Tahoe 1.8.x and 1.9.x
 - Brian Warner - began writing the introducerless mode
 - David Stainton - made introducerless mode work
 - Leif Ryge - ported to present, added introducerless config generator
 - Anonymous - wrote tests and fixed bugs

The history of this patch is documented at
https://tahoe-lafs.org/trac/tahoe-lafs/ticket/68 and
https://tahoe-lafs.org/trac/tahoe-lafs/ticket/467

This commit is almost on top of current upstream master; it contains just one
ancestor which is not yet merged (but does have an open pull request).

This commit will likely end up in truckee but not master; I will add more tests
and rebase again before submitting this there.