quick truckee update
truckee is quite out of date with master now; sorry for that!

this commit contains a quick hack which seems to make it work with torsocks2.
unfortunately there seems to be some problem still where torsocks2 causes tahoe
to max out the CPU for a little while; I'm investigating that.

this also has a fix to make things work with foolscap-0.8 (and not the earlier
versions anymore).