revert an old commit to fix attribution (vanity)
This approximately reverts a2d724aab7bf6dd2f7ffde37cb21c570ba773bf8 (modulo the
one line that daira added) so that my contributions will be properly attributed
by git blame etc. Another version of this commit exists on master and is
probably in a working state; this commit is likely not working and was created
to fix the attribution from the perspective of the truckee branch.

The next commit after this one should re-add the 1973 feature branch and leave
truckee in a working state.